+++ backup_scrapbox/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from ._config import Config
+from ._main import backup_scrapbox
