+++ backup_scrapbox/__main__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from ._main import main
+
+if __name__ == '__main__':
+    main()
